requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.5
django-bootstrap4==22.1
django-filter==22.1
Pillow==9.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
